Move most functionality into new file helpers.py
